# Optimized @async_unsafe.