@adamchainz 2021-09-29
Before:

```
In [1]: from django.utils.asyncio import async_unsafe

In [2]: ai = async_unsafe(int)

In [3]: %timeit ai()
1.37 µs ± 13.6 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

After:

```
In [1]: from django.utils.asyncio import async_unsafe

In [2]: ai = async_unsafe(int)

In [3]: %timeit ai()
395 ns ± 1.68 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

Calling `int()` takes ~50ns, so we can take that from each number, giving ~1300ns before and ~350ns after, ~4 times faster.

A small boost but it multiplies as `async_unsafe` decorated functions are called several times per query.