@adamchainz 2021-09-30
> Are the import changes related to the optimisation? 🤔

They're another micro-optimization to require fewer attribute accesses.