@tim-mccurrach 2021-09-30
Whilst we're micro-optimising could we use `_get_running_loop` to shave a few more microseconds off. Then `inner` would become:
```
def inner(*args, **kwargs):
    if not _get_running_loop():
        return func(*args, **kwargs)
    if not os.environ.get('DJANGO_ALLOW_ASYNC_UNSAFE'):
        raise ValueError(message)
```
Or would the use of the private function be too subject-to-change for us to use?
Unfortunately, I'm not able to test this for time right now, but I suspect it would make it marginally quicker (famous last words...).