@tim-mccurrach 2021-09-30
@carltongibson 
Although having said that, quickly testing just now, it does seem to cut the time in half again:
```
>>> timeit.repeat("ai()",repeat=7, number=100000, setup="from __main__ import ai")
[0.0292515020000792, 0.023990991999653488, 0.020490372999574902, 0.020217061000039394, 0.0210146330000498, 0.020252889999937906, 0.020329448000211414]
>>> timeit.repeat("ai2()",repeat=7, number=100000, setup="from __main__ import ai2")
[0.05905059799988521, 0.04216317100008382, 0.042790552000042226, 0.04461314699983632, 0.04397789999984525, 0.04326298099977066, 0.04322846100012612]
```
(ai2 is the function as it stands in this PR). 

It's also (perhaps more importantly) more readable (IMO).

But I agree, it may well be a bridge too far 🤷