@tim-mccurrach 2021-09-30
> @tim-mccurrach given that Django 4.1 will support 4 or 5 versions of Python using private APIs is not a good idea.

👍  Yes, that seems sensible.